fix: invalid feature flag usage
