+++ packages/rs-dpp/src/state_transition/state_transitions/document/batch_transition/methods/mod.rs
@@ -20,9 +20,6 @@ use crate::prelude::UserFeeIncrease;
 use crate::state_transition::batch_transition::batched_transition::BatchedTransition;
 use crate::state_transition::batch_transition::methods::v0::DocumentsBatchTransitionMethodsV0;
 use crate::state_transition::batch_transition::methods::v1::DocumentsBatchTransitionMethodsV1;
-use crate::state_transition::batch_transition::token_transfer_transition::{
-    PrivateEncryptedNote, SharedEncryptedNote,
-};
 use crate::state_transition::batch_transition::BatchTransition;
 #[cfg(feature = "state-transition-signing")]
 use crate::state_transition::batch_transition::{BatchTransitionV0, BatchTransitionV1};
@@ -30,6 +27,7 @@ use crate::state_transition::batch_transition::{BatchTransitionV0, BatchTransiti
 use crate::state_transition::StateTransition;
 #[cfg(feature = "state-transition-signing")]
 use crate::tokens::emergency_action::TokenEmergencyAction;
+use crate::tokens::{PrivateEncryptedNote, SharedEncryptedNote};
 use crate::ProtocolError;
 #[cfg(feature = "state-transition-signing")]
 use platform_value::Identifier;
