+++ packages/rs-dpp/src/tokens/mod.rs
@@ -1,4 +1,7 @@
 use crate::data_contract::TokenContractPosition;
+use crate::prelude::{
+    DerivationEncryptionKeyIndex, RecipientKeyIndex, RootEncryptionKeyIndex, SenderKeyIndex,
+};
 use crate::util::hash::hash_double;
 
 pub mod allowed_currency;
@@ -7,6 +10,13 @@ pub mod errors;
 pub mod info;
 pub mod status;
 pub mod token_event;
+
+pub type SharedEncryptedNote = (SenderKeyIndex, RecipientKeyIndex, Vec<u8>);
+pub type PrivateEncryptedNote = (
+    RootEncryptionKeyIndex,
+    DerivationEncryptionKeyIndex,
+    Vec<u8>,
+);
 
 pub fn calculate_token_id(contract_id: &[u8; 32], token_pos: TokenContractPosition) -> [u8; 32] {
     let mut bytes = b"dash_token".to_vec();
