+++ packages/rs-drive/src/state_transition_action/batch/batched_transition/token_transition/token_transfer_transition_action/mod.rs
@@ -3,7 +3,7 @@ use derive_more::From;
 use crate::state_transition_action::batch::batched_transition::token_transition::token_base_transition_action::TokenBaseTransitionAction;
 use dpp::identifier::Identifier;
 use dpp::prelude::{DerivationEncryptionKeyIndex, RootEncryptionKeyIndex};
-use dpp::state_transition::batch_transition::token_transfer_transition::{PrivateEncryptedNote, SharedEncryptedNote};
+use dpp::tokens::{PrivateEncryptedNote, SharedEncryptedNote};
 
 /// transformer module
 pub mod transformer;
