+++ packages/rs-drive/src/state_transition_action/batch/batched_transition/token_transition/token_transfer_transition_action/v0/mod.rs
@@ -6,7 +6,7 @@ use dpp::identifier::Identifier;
 use dpp::prelude::{
     DerivationEncryptionKeyIndex, IdentityNonce, RootEncryptionKeyIndex,
 };
-use dpp::state_transition::batch_transition::token_transfer_transition::{PrivateEncryptedNote, SharedEncryptedNote};
+use dpp::tokens::{PrivateEncryptedNote, SharedEncryptedNote};
 use crate::drive::contract::DataContractFetchInfo;
 use crate::state_transition_action::batch::batched_transition::token_transition::token_base_transition_action::{
     TokenBaseTransitionAction, TokenBaseTransitionActionAccessorsV0,
